% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRIComBat.R
\name{fMRIComBat}
\alias{fMRIComBat}
\title{fMRI ComBat}
\usage{
fMRIComBat(BOLD, batch, nreg = NULL, TR = NULL, hpf = 0.01, scale = TRUE)
}
\arguments{
\item{BOLD}{file paths to the BOLD data as CIFTI, NIFTI, or GIFTI files.}

\item{batch}{Key for the batch of each entry in \code{BOLD}.}

\item{...}{Description}
}
\value{
The value
}
\description{
Batch correction for fMRI data
}
