% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRIComBat_preproc.R
\name{fMRIComBat_preproc}
\alias{fMRIComBat_preproc}
\title{Pre-process BOLD data for ComBat modelling}
\usage{
fMRIComBat_preproc(BOLD, return_value = c("BOLD", "params"), nreg = NULL)
}
\arguments{
\item{BOLD}{The \eqn{V \times T} fMRI data.}

\item{return_value}{Return the \code{BOLD} data after
pre-processing (\code{"BOLD"}, default) or the measurements needed
to reconstruct it (\code{"params"})?}
}
\description{
Pre-process the BOLD data in preparation for ComBat modelling.
}
