% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preclean}
\alias{preclean}
\title{Pre-clean the data}
\usage{
preclean(X, nreg = NULL, TR, hpf = 0.01, lpf = NULL)
}
\arguments{
\item{X}{The \eqn{T \times V} data.}
}
\value{
The value
}
\description{
Perform nuisance regression and temporal filtering.
}
